[PeerRenames][Part 3] move /list/single to /chooser/single
Summary: As the title says, also renames the package from list to
chooser
